fix(pr-1271): resolve 16 page-2 review comments
ParameterCount int-overflow fixes (cast first term to long so the running
sum widens to 64-bit before reaching ToFlatVectorSize, preventing wraparound
on multi-billion-parameter configs):
- MixtureOfMambaLayer.cs, S5Layer.cs, GatedLinearAttentionLayer.cs
- HybridBlockScheduler.cs, HyenaLayer.cs (foreach long accumulator)
- InteractingLayer.cs, FourierNeuralOperator.cs (both FNO + FourierLayer)
- SymmetricProjector.cs (introduce ComputeParameterCountLong)
- VideoCLIPNeuralNetwork.cs (long count + (long) on matrix-element products)
- PointNetPlusPlus.cs (foreach long accumulator)

Allocation / lifetime correctness:
- SubpixelConvolutionalLayer.cs: InitializeWeights writes IN PLACE into the
  AllocateLazyWeight-registered _kernels tensor instead of replacing the
  field reference, preserving streaming-pool registration.
- LocallyConnectedLayer.cs: SetParameters copies into existing tensor
  storage rather than `Tensor<T>.FromVector(...)` so the engine's persistent-
  tensor registry doesn't follow stale references.
- TimeMoEBlockLayer.cs: switch from sub.GetParameters().Length (which
  materialises a multi-billion-parameter Vector<T> just to read its length)
  to (int)sub.ParameterCount. Sublayer types maintain ParameterCount ==
  GetParameters().Length once IsShapeResolved is true.

API / contract correctness:
- IAiModelBuilder.cs: extract ConfigureWeightStreaming into a new companion
  interface IWeightStreamingCapableBuilder<T,TInput,TOutput> that extends
  IAiModelBuilder, so introducing the method does not break external
  implementers of IAiModelBuilder. AiModelBuilder<T,TInput,TOutput> now
  implements both. AiModelResult cref updated.
- DecoderLayer.cs: throw when SetParameters is called before _feedForward2
  is constructed instead of silently turning the slice into a no-op (which
  would misalign the trailing norm slices).
- GRULayer.cs: validate the inferred inputSize round-trips to the exact
  parameter count before calling ResolveFromShape, rejecting malformed
  vectors that happen to land on a 3*hiddenSize multiple.
- SeparableConvolutionalLayer.cs: pass the rank-3 [H, W, C] per-sample
  shape to ResolveFromShape; the previous rank-4 form put candidateInputDepth
  in the channel slot only by accident.
- TimeEmbeddingLayer.cs: GetParameterGradients always returns a Vector<T>
  of ParameterCount length, zero-filling slots whose gradient tensors are
  null, so callers see a stable shape regardless of partial gradient state.

Streaming guards:
- NeuralNetworkBase.cs RefreshWeightRegistry: skip GetExtraTrainableTensors
  entries with StreamingPoolHandle >= 0 so re-registration after the model
  is already streaming doesn't trigger the dropped-storage AsSpan() path.
- NeuralNetworkBase.cs ResetWeightStreamingForTests: surface the original
  exception with a wrapping note instead of swallowing — process-wide
  singleton corruption shouldn't be silent.

Silent-zero-gradient fixes:
- OnlineLearningModelBase.ComputeGradients: throw NotSupportedException by
  default instead of returning a zero vector that lets gradient-based
  optimizers proceed with no real signal.
- SurvivalModelBase.ComputeGradients: same.

Test correctness:
- WeightStreamingEndToEndTests.cs Streaming_TightPoolBudget: add positive
  eviction signal (resident > 0 || EvictionCount > 0) so the test fails
  when the eviction path is inert, not just when it exceeds budget.

AiModelBuilder report:
- Capture exception reason in WeightStreamingReport.CountersUnavailableReason
  instead of silently swallowing on counter-read failure (page-1 comment).
- WeightStreamingReport.cs: add CountersUnavailableReason field.
- PatchEmbeddingLayer.cs: route SetParameters allocations through
  AllocateLazyWeight; add _paramsLoadedViaSetParameters flag so OnFirstForward
  preserves caller-loaded biases (page-1 comments).
- SparseLinearLayer.cs Deserialize: reconstruct _weights with the saved
  sparsity pattern when nnz != current NonZeroCount, validate index ranges
  (page-1 comment).
- DenseLayer.cs Deserialize: clarify comment to match implementation
  (page-1 comment).

+++ src/NeuralNetworks/Layers/SSM/GatedLinearAttentionLayer.cs
@@ -130,7 +130,10 @@ internal partial class GatedLinearAttentionLayer<T> : LayerBase<T>
     /// Gets the total number of trainable parameters.
     /// </summary>
     public override long ParameterCount =>
-        _queryWeights.Length + _keyWeights.Length + _valueWeights.Length +
+        // Cast the first term to long so the running sum widens to 64-bit
+        // and never wraps before reaching ToFlatVectorSize on multi-billion-
+        // parameter linear-attention configs.
+        (long)_queryWeights.Length + _keyWeights.Length + _valueWeights.Length +
         _gateWeights.Length + _gateBias.Length +
         _outputWeights.Length + _outputBias.Length;
