+++ src/NeuralNetworks/Layers/SSM/HybridBlockScheduler.cs
@@ -147,9 +147,12 @@ public partial class HybridBlockScheduler<T> : LayerBase<T>
     {
         get
         {
-            int count = 0;
+            // Accumulate in long so multi-block schedulers (e.g. 64+ Mamba/SSM
+            // blocks each with multi-billion-parameter state) don't wrap
+            // before reaching ToFlatVectorSize.
+            long count = 0;
             foreach (var block in _blocks)
-                count += (int)block.ParameterCount;
+                count += block.ParameterCount;
             foreach (var gamma in _normGammas)
                 count += gamma.Length;
             foreach (var beta in _normBetas)
