+++ src/NeuralNetworks/Layers/SSM/HyenaLayer.cs
@@ -153,7 +153,11 @@ public partial class HyenaLayer<T> : LayerBase<T>
     {
         get
         {
-            int count = 0;
+            // Accumulate in long: a high-order Hyena (order >= 2) over a
+            // long sequence has filter networks whose weight tensors
+            // individually approach int.MaxValue; summing in int can wrap
+            // before ToFlatVectorSize sees the value.
+            long count = 0;
 
             // Input projections: (order + 1) x (weights + bias)
             for (int i = 0; i <= _order; i++)
